requirements.txt
# TikTok Recipe Newsletter Dependencies

# TikTok API wrapper
TikTokApi==6.3.1
playwright==1.40.0

# AI for recipe extraction
anthropic==0.39.0

# Email service
resend==2.0.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
